Refactor component names for readability
